feat: move streams into src folder
